Update: New class.
* graftm/create.py: Move update from functionality from here ...
* graftm/update.py: ... to here.
* graftm/run.py: update references to Update.
